+++ backend/api/query.py
@@ -14,9 +14,11 @@ def query_db():
     
     data = request.get_json()
     query = data.get("query", "")
+    access_token = data.get("access_token", "")
     print("Query:", query)
     embedding = embed_texts(query)
-    results = query_vectors(embedding)
+    results = query_vectors(embedding, access_token)
+    
     top_ids = results["ids"][0]
     top_texts = results["documents"][0]
     top_distances = results["distances"][0]
