+++ backend/models/recipe.py
@@ -1,3 +1,5 @@
+from extensions import db
+
 class Recipe(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer, db.ForeignKey("user.id"))
