+++ backend/requirements.txt
@@ -15,3 +15,4 @@ google-generativeai
 Flask-SQLAlchemy
 google-auth
 flask-jwt-extended
+Flask-Migrate
